+++ src/metric.hpp
@@ -93,4 +93,35 @@ private:
     friend class MetricFamily<MetricHistogram>;
 };
 
+// Increments upon destruction, however can be disabled to do so.
+class MetricCounterGuard {
+    bool active = true;
+    MetricCounter* metric;
+
+public:
+    MetricCounterGuard(MetricCounter* metric) :
+        metric(metric) {
+    }
+    void disable() { active = false; }
+    ~MetricCounterGuard() {
+        if (active) {
+            INCREMENT_IF_ENABLED(metric);
+        }
+    }
+};
+
+// Increments upon construction, decrements upon destruction.
+class MetricGaugeGuard {
+    MetricGauge* metric;
+
+public:
+    MetricGaugeGuard(MetricGauge* metric) :
+        metric(metric) {
+        INCREMENT_IF_ENABLED(metric);
+    }
+    ~MetricGaugeGuard() {
+        DECREMENT_IF_ENABLED(metric);
+    }
+};
+
 }  // namespace ovms
